+++ ai_portfolio/src/components/three/ThreeCanvas.tsx
@@ -1,5 +1,4 @@
 'use client';
-import React, { useState, useRef, useEffect, useCallback } from 'react';
 import { Canvas } from '@react-three/fiber';
 import { Scroll, ScrollControls } from '@react-three/drei';
 import ScrollImg from '@/components/templetes/scroll-img';
@@ -10,107 +9,11 @@ import TopSection from '../sections/TopSection';
 import MessageSection from '../sections/Message';
 import ExperienceSection from '../sections/ExperienceSection';
 import { useVisibleSections } from '@/hooks';
+import ItSection from '../sections/ItSection';
 
 const ThreeCanvas = () => {
-  const [currentSlide, setCurrentSlide] = useState(0);
-  const carouselRef = useRef<HTMLDivElement>(null);
-  const totalSlides = 3;
+  const { setVisibleSections, isVisible } = useVisibleSections();
 
-  // /* 表示されているセクションのIDを格納するSet ['Hello', 'Profile'...ets] */
-  // const [visibleSections, setVisibleSections] = useState<Set<string>>(
-  //   new Set()
-  // );
-  // /**
-  //  * セクションの表示状態を判定
-  //  * @param sectionId - セクションのID{"hello", "profile"...}
-  //  * @returns セクションが表示されているかどうか真偽値
-  //  */
-  // const isVisible = (sectionId: string) => visibleSections.has(sectionId);
-  // console.log(visibleSections);
-
-  /**
-   * セクションの表示状態を管理するカスタムフック
-   */
-  const { setVisibleSections, isVisible, visibleSections } =
-    useVisibleSections();
-
-  // 動的にスライド幅を計算する関数
-  const getSlideWidth = useCallback(() => {
-    if (!carouselRef.current) return 400;
-
-    const container = carouselRef.current;
-    const containerWidth = (container as HTMLElement).offsetWidth; // コンテナ幅に基づいてスライド幅を計算（gap分も考慮）
-    return containerWidth;
-  }, []);
-
-  // ドットの状態を更新する関数
-  const updateDots = useCallback(() => {
-    if (!carouselRef.current) return;
-
-    const container = carouselRef.current;
-    const scrollLeft = (container as HTMLElement).scrollLeft;
-    const slideWidth = getSlideWidth();
-    const newSlide = Math.round(scrollLeft / slideWidth);
-    setCurrentSlide(Math.max(0, Math.min(newSlide, totalSlides - 1)));
-  }, [getSlideWidth, totalSlides]);
-
-  // 指定されたスライドに移動する関数
-  const goToSlide = useCallback(
-    (slideIndex: number) => {
-      if (!carouselRef.current) return;
-
-      const slideWidth = getSlideWidth();
-      carouselRef.current.scrollTo({
-        left: slideIndex * slideWidth,
-        behavior: 'smooth',
-      });
-      setCurrentSlide(slideIndex);
-    },
-    [getSlideWidth]
-  );
-
-  // 前のスライドに移動
-  const prevSlide = useCallback(() => {
-    const newSlide = Math.max(0, currentSlide - 1);
-    goToSlide(newSlide);
-  }, [currentSlide, goToSlide]);
-
-  // 次のスライドに移動
-  const nextSlide = useCallback(() => {
-    const newSlide = Math.min(totalSlides - 1, currentSlide + 1);
-    goToSlide(newSlide);
-  }, [currentSlide, totalSlides, goToSlide]);
-
-  // リサイズ時の処理
-  const handleResize = useCallback(() => {
-    // リサイズ後に現在のスライドの位置を調整
-    setTimeout(() => {
-      goToSlide(currentSlide);
-    }, 100);
-  }, [currentSlide, goToSlide]);
-
-  // スクロール終了時のイベントハンドラー
-  useEffect(() => {
-    const container = carouselRef.current;
-    if (!container) return;
-
-    let scrollTimeout: NodeJS.Timeout | undefined;
-    const handleScroll = () => {
-      clearTimeout(scrollTimeout);
-      scrollTimeout = setTimeout(updateDots, 150);
-    };
-
-    container.addEventListener('scroll', handleScroll, { passive: true });
-
-    // リサイズイベントリスナーを追加
-    window.addEventListener('resize', handleResize);
-
-    return () => {
-      container.removeEventListener('scroll', handleScroll);
-      window.removeEventListener('resize', handleResize);
-      clearTimeout(scrollTimeout);
-    };
-  }, [updateDots, handleResize]);
   return (
     <>
       <Canvas
@@ -133,486 +36,7 @@ const ThreeCanvas = () => {
             {/* Previous Experience */}
             <ExperienceSection isVisible={isVisible} />
             {/* IT */}
-            <section
-              id='it'
-              className='bg-[#bfdad1] relative top-[355vh] w-screen h-screen p-8 mt-8 py-6'
-            >
-              <div className='double-line-header flex items-center w-full relative my-10 mx-0'></div>
-              <div className='container flex justify-left gap-4 flex-wrap justify-between m-auto'>
-                {/* IT左側 */}
-                <div className='w-6/12 flex flex-col'>
-                  <h3 className='text-3xl mb-1 pb-12'>
-                    IT業界<span className='jp text-sm ml-2'> 2018</span>
-                  </h3>
-                  <p className='leading-loose'>
-                    ただIT業界と言っても右も左も分からない状態； <br />
-                    なので一先ずスクールに通い友達のネイルサイトを作成し、転職活動に挑みました！
-                    <br />
-                    有り難いことに拾って頂きSESとして未熟な私にも様々な現場で経験が出来、本当に感謝をしています。
-                    <br />
-                    契約の関係上、作成したサイトは転載する事が出来ませんがその現場での業務を記載しましたので見て頂けると幸いです。　
-                  </p>
-                </div>
-
-                {/* IT右側 - カルーセル風スライダー */}
-                <div className='w-5/12 mt-3.5'>
-                  {/* カルーセルコンテナ */}
-                  <div className='relative max-w-full mx-auto'>
-                    {/* 左矢印ボタン */}
-                    <button
-                      onClick={prevSlide}
-                      disabled={currentSlide === 0}
-                      className={`absolute left-[-60px] top-1/2 transform -translate-y-1/2 z-30 w-12 h-12 rounded-full shadow-lg hover:shadow-xl transition-all duration-200 flex items-center justify-center group ${
-                        currentSlide === 0
-                          ? 'bg-gray-300 cursor-not-allowed'
-                          : 'bg-white/90 hover:bg-white'
-                      }`}
-                      aria-label='前のプロジェクトを見る'
-                    >
-                      <div className='flex items-center gap-[-2px]'>
-                        <div
-                          className={`w-3 h-3 border-l-2 border-b-2 transform rotate-45 transition-colors ${
-                            currentSlide === 0
-                              ? 'border-gray-500'
-                              : 'border-gray-600 group-hover:border-gray-800'
-                          }`}
-                        ></div>
-                        <div
-                          className={`w-3 h-3 border-l-2 border-b-2 transform rotate-45 -ml-1 transition-colors ${
-                            currentSlide === 0
-                              ? 'border-gray-500'
-                              : 'border-gray-600 group-hover:border-gray-800'
-                          }`}
-                        ></div>
-                      </div>
-                    </button>
-
-                    {/* 右矢印ボタン */}
-                    <button
-                      onClick={nextSlide}
-                      disabled={currentSlide === totalSlides - 1}
-                      className={`absolute right-[-60px] top-1/2 transform -translate-y-1/2 z-30 w-12 h-12 rounded-full shadow-lg hover:shadow-xl transition-all duration-200 flex items-center justify-center group ${
-                        currentSlide === totalSlides - 1
-                          ? 'bg-gray-300 cursor-not-allowed'
-                          : 'bg-white/90 hover:bg-white'
-                      }`}
-                      aria-label='次のプロジェクトを見る'
-                    >
-                      <div className='flex items-center gap-[-2px]'>
-                        <div
-                          className={`w-3 h-3 border-r-2 border-b-2 transform -rotate-45 transition-colors ${
-                            currentSlide === totalSlides - 1
-                              ? 'border-gray-500'
-                              : 'border-gray-600 group-hover:border-gray-800'
-                          }`}
-                        ></div>
-                        <div
-                          className={`w-3 h-3 border-r-2 border-b-2 transform -rotate-45 -ml-1 transition-colors ${
-                            currentSlide === totalSlides - 1
-                              ? 'border-gray-500'
-                              : 'border-gray-600 group-hover:border-gray-800'
-                          }`}
-                        ></div>
-                      </div>
-                    </button>
-
-                    {/* スライドコンテナ */}
-                    <div
-                      ref={carouselRef}
-                      className='overflow-x-auto scroll-smooth'
-                      style={{
-                        scrollSnapType: 'x mandatory',
-                        scrollbarWidth: 'none' /* Firefox */,
-                        msOverflowStyle: 'none' /* Internet Explorer 10+ */,
-                      }}
-                    >
-                      <div className='flex'>
-                        {/* プロジェクト1: ショッピングサイト */}
-                        <div
-                          className='w-full flex-shrink-0 px-2'
-                          style={{ scrollSnapAlign: 'start' }}
-                        >
-                          <div className='bg-white p-5 border-double shadow-[0_0_0_15px_#eaf4f4] border-[#a0b4b4] border-4'>
-                            <div className='max-w-2xl mx-auto bg-white'>
-                              <table className='w-full'>
-                                <thead>
-                                  <tr>
-                                    <th className='p-4 text-lg text-center jp'>
-                                      ショッピングサイト
-                                    </th>
-                                  </tr>
-                                </thead>
-                                <tbody className='text-sm'>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <div className='space-y-1.5 jp'>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-8'>参画期間</span>
-                                          <span>自社/出社/3か月</span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-16'>規模</span>
-                                          <span>ディレクター1名、FE5名</span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-6'>
-                                            担当フェーズ
-                                          </span>
-                                          <span>開発</span>
-                                        </div>
-                                        <div>
-                                          <div className='flex items-start mb-2'>
-                                            <span className='text-black mr-2'>
-                                              ◆
-                                            </span>
-                                            <span className='font-semibold'>
-                                              業務内容
-                                            </span>
-                                          </div>
-                                          <div className='ml-6'>
-                                            <span className='text-black mr-2'>
-                                              •
-                                            </span>
-                                            <span>会員ページの改修作業</span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='px-6 py-4'>
-                                      <div className='border-t-2 border-dashed border-gray-400 pt-4'>
-                                        <h3 className='font-bold text-lg mb-4 jp'>
-                                          開発環境
-                                        </h3>
-                                        <div className='space-y-1.5 jp'>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【言語】
-                                            </span>
-                                            <span className='ml-2'>
-                                              HTML/CSS/PHP
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【OS】
-                                            </span>
-                                            <span className='ml-4'>
-                                              Windows
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【FW】
-                                            </span>
-                                            <span className='ml-4'>
-                                              Laravel
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【ツール】
-                                            </span>
-                                            <span className='ml-2'>
-                                              Chatwork/GitHub
-                                            </span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <p className='leading-relaxed jp'>
-                                        初の実務だった為、まずは先輩方から基本作法を学びながらページの改修作業を行いました。
-                                      </p>
-                                    </td>
-                                  </tr>
-                                </tbody>
-                              </table>
-                            </div>
-                          </div>
-                        </div>
-
-                        {/* プロジェクト2: 企業サイト */}
-                        <div
-                          className='w-full flex-shrink-0 px-2'
-                          style={{ scrollSnapAlign: 'start' }}
-                        >
-                          <div className='bg-white p-5 border-double shadow-[0_0_0_15px_#eaf4f4] border-[#a0b4b4] border-4'>
-                            <div className='max-w-2xl mx-auto bg-white'>
-                              <table className='w-full'>
-                                <thead>
-                                  <tr>
-                                    <th className='p-4 text-lg text-center jp'>
-                                      企業サイト
-                                    </th>
-                                  </tr>
-                                </thead>
-                                <tbody className='text-sm'>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <div className='space-y-1.5 jp'>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-8'>参画期間</span>
-                                          <span>客先常駐/6か月</span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-16'>規模</span>
-                                          <span>PM1名、FE3名、BE2名</span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-6'>
-                                            担当フェーズ
-                                          </span>
-                                          <span>設計・開発・テスト</span>
-                                        </div>
-                                        <div>
-                                          <div className='flex items-start mb-2'>
-                                            <span className='text-black mr-2'>
-                                              ◆
-                                            </span>
-                                            <span className='font-semibold'>
-                                              業務内容
-                                            </span>
-                                          </div>
-                                          <div className='ml-6'>
-                                            <span className='text-black mr-2'>
-                                              •
-                                            </span>
-                                            <span>
-                                              コーポレートサイトの新規構築
-                                            </span>
-                                          </div>
-                                          <div className='ml-6'>
-                                            <span className='text-black mr-2'>
-                                              •
-                                            </span>
-                                            <span>レスポンシブ対応</span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='px-6 py-4'>
-                                      <div className='border-t-2 border-dashed border-gray-400 pt-4'>
-                                        <h3 className='font-bold text-lg mb-4 jp'>
-                                          開発環境
-                                        </h3>
-                                        <div className='space-y-1.5 jp'>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【言語】
-                                            </span>
-                                            <span className='ml-2'>
-                                              HTML/CSS/JavaScript
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【OS】
-                                            </span>
-                                            <span className='ml-4'>Mac</span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【FW】
-                                            </span>
-                                            <span className='ml-4'>Vue.js</span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【ツール】
-                                            </span>
-                                            <span className='ml-2'>
-                                              Slack/GitLab
-                                            </span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <p className='leading-relaxed jp'>
-                                        フロントエンド開発を中心に、デザインからコーディングまで幅広く担当しました。
-                                      </p>
-                                    </td>
-                                  </tr>
-                                </tbody>
-                              </table>
-                            </div>
-                          </div>
-                        </div>
-
-                        {/* プロジェクト3: データ可視化サイト */}
-                        <div
-                          className='w-full flex-shrink-0 px-2'
-                          style={{ scrollSnapAlign: 'start' }}
-                        >
-                          <div className='bg-white p-5 border-double shadow-[0_0_0_15px_#eaf4f4] border-[#a0b4b4] border-4'>
-                            <div className='max-w-2xl mx-auto bg-white'>
-                              <table className='w-full'>
-                                <thead>
-                                  <tr>
-                                    <th className='p-4 text-lg text-center jp'>
-                                      データ可視化サイト
-                                    </th>
-                                  </tr>
-                                </thead>
-                                <tbody className='text-sm'>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <div className='space-y-1.5 jp'>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-8'>参画期間</span>
-                                          <span>リモート/1年</span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-16'>規模</span>
-                                          <span>
-                                            PL1名、FE4名、デザイナー2名
-                                          </span>
-                                        </div>
-                                        <div className='flex items-start'>
-                                          <span className='mr-2'>◆</span>
-                                          <span className='mr-6'>
-                                            担当フェーズ
-                                          </span>
-                                          <span>要件定義・設計・開発</span>
-                                        </div>
-                                        <div>
-                                          <div className='flex items-start mb-2'>
-                                            <span className='text-black mr-2'>
-                                              ◆
-                                            </span>
-                                            <span className='font-semibold'>
-                                              業務内容
-                                            </span>
-                                          </div>
-                                          <div className='ml-6'>
-                                            <span className='text-black mr-2'>
-                                              •
-                                            </span>
-                                            <span>D3.jsを使ったグラフ作成</span>
-                                          </div>
-                                          <div className='ml-6'>
-                                            <span className='text-black mr-2'>
-                                              •
-                                            </span>
-                                            <span>
-                                              インタラクティブな可視化
-                                            </span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='px-6 py-4'>
-                                      <div className='border-t-2 border-dashed border-gray-400 pt-4'>
-                                        <h3 className='font-bold text-lg mb-4 jp'>
-                                          開発環境
-                                        </h3>
-                                        <div className='space-y-1.5 jp'>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【言語】
-                                            </span>
-                                            <span className='ml-2'>
-                                              TypeScript/HTML/CSS
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【OS】
-                                            </span>
-                                            <span className='ml-4'>Mac</span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【FW】
-                                            </span>
-                                            <span className='ml-4'>
-                                              Next.js
-                                            </span>
-                                          </div>
-                                          <div>
-                                            <span className='font-semibold'>
-                                              【ツール】
-                                            </span>
-                                            <span className='ml-2'>
-                                              GitHub/Figma
-                                            </span>
-                                          </div>
-                                        </div>
-                                      </div>
-                                    </td>
-                                  </tr>
-                                  <tr>
-                                    <td className='p-6'>
-                                      <p className='leading-relaxed jp'>
-                                        データビジュアライゼーションライブラリを活用し、ユーザーにとって分かりやすいグラフィカルな表現を実現しました。
-                                      </p>
-                                    </td>
-                                  </tr>
-                                </tbody>
-                              </table>
-                            </div>
-                          </div>
-                        </div>
-                      </div>
-                    </div>
-
-                    {/* スライド番号表示 */}
-                    <div className='flex justify-center items-center mt-6 gap-4'>
-                      {/* ページネーション */}
-                      <div className='text-sm text-gray-600 font-medium bg-white/80 px-3 py-1 rounded-full'>
-                        {currentSlide + 1} / {totalSlides}
-                      </div>
-
-                      {/* インジケーター（ドット） */}
-                      <div className='flex gap-2'>
-                        {[...Array(totalSlides)].map((_, index) => (
-                          <button
-                            key={index}
-                            className={`w-3 h-3 rounded-full transition-colors ${
-                              currentSlide === index
-                                ? 'bg-gray-800'
-                                : 'bg-gray-400 hover:bg-gray-600'
-                            }`}
-                            onClick={() => goToSlide(index)}
-                            aria-label={`${index + 1}番目のプロジェクトを表示`}
-                          />
-                        ))}
-                      </div>
-                    </div>
-                  </div>
-                </div>
-              </div>
-
-              <div className='absolute opacity-80 bottom-5 left-10'>
-                {/* eslint-disable-next-line @next/next/no-img-element */}
-                <img
-                  src='../assets/images/nail_1.png'
-                  alt='ネイル'
-                  className=''
-                />
-              </div>
-              <div className='absolute left-1/2 bottom-0 -translate-x-1/2 -translate-y-1/2 text-[#eaf4f4] text-9xl md:text-[10rem] font-bold whitespace-nowrap pointer-events-none select-none'>
-                Previous Experience
-              </div>
-            </section>
+            <ItSection isVisible={isVisible} />
             {/* スキル */}
             <section
               id='skill'
